Migrate fjs/dev to .f.mjs; fix CHANGELOG Entity wording
Splits Module, ModuleMap, and LoadModuleOperations into a sibling
types.ts. proof.f.ts migrates alongside it since its only runtime
dependency is the module itself; the embedded proof export inside
module.f.ts (allFiles/loadModuleMap coverage) moves over unchanged.

Updates the one dependent, fjs/emergent_testing, and stale doc
references and line-number citations across fjs/dev/todo,
fjs/effects/todo, fjs/effects/node/todo, and fjs/emergent_testing/todo.

Also fixes the previous commit's CHANGELOG entry for
fjs/effects/node/virtual: Entity was a publicly exported type on main,
not internal, so the entry undersold the change by calling it "the
internal _Entity type". Reworded to "narrowing the previously-exported
Entity to the internal _Entity", per review feedback from
o2alexanderfedin.
