+++ fjs/emergent_testing/proof.f.ts
@@ -12,7 +12,7 @@ import {
 } from './module.f.ts'
 import { run as mockRun } from '../effects/mock/module.f.mjs'
 import type { All, Await, Test, TestContext } from '../effects/node/types.ts'
-import { shouldLoad } from '../dev/module.f.ts'
+import { shouldLoad } from '../dev/module.f.mjs'
 import { parse as parseJson } from '../media/json/module.f.ts'
 import { array, number as rttiNumber, or, string as rttiString } from '../types/rtti/module.f.mjs'
 import { parse as rttiParse } from '../types/rtti/parse/module.f.mjs'
